only run uninstall functions if dependencies are available

+++ plugins/woocommerce-admin/uninstall.php
@@ -9,5 +9,7 @@ defined( 'WP_UNINSTALL_PLUGIN' ) || exit;
 
 require_once dirname( __FILE__ ) . '/woocommerce-admin.php';
 
-WC_Admin_Feature_Plugin::instance()->includes();
-WC_Admin_Install::delete_table_data();
+if ( WC_Admin_Feature_Plugin::instance()->check_dependencies() ) {
+	WC_Admin_Feature_Plugin::instance()->includes();
+	WC_Admin_Install::delete_table_data();
+}
